split game box renderer into multiple files
